backend/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
werkzeug>=2.3.0 